Utest per build intero corso

+++ Demo/Test/SBATCH/SBATCH/SBatch170-Exceptions/src/main/java/corso/spring/batch/demo/modulo/exceptions/job/chuncks/IntegerReader.java
@@ -22,7 +22,7 @@ public class IntegerReader extends AbstractChunkOrientedDemo  implements ItemRea
 		log.info("READER Summary: "+this.stepExecution.getSummary());	
 				
 		Integer numberToSave=numberService.produceNumber();
-		log.info(" Retrieved: "+numberToSave.intValue());	
+		log.info(" Retrieved: "+numberToSave);	
 		
 		//executeDangerousOperation(numberToSave);		
 		numberService.ackNumber();
